// src/enums/httpStatus.enum.ts
export enum HttpStatus {
	// 2xx Success
	OK = 200,
	CREATED = 201,
	NO_CONTENT = 204,
	ACCEPTED = 202,

	// 4xx Client Errors
	BAD_REQUEST = 400,
	UNAUTHORIZED = 401,
	FORBIDDEN = 403,
	NOT_FOUND = 404,
	CONFLICT = 409,
	TOO_MANY_REQUESTS = 429,

	// 5xx Server Errors
	INTERNAL_SERVER_ERROR = 500,
	SERVICE_UNAVAILABLE = 503,
}